Make back-end changes to support open game changes
Add myTableId, ordinal, and a list of tables they're viewing to the
OpenGameForUser class, which allows the front-end to show a bunch more
data.  Adjust the way the repository stores the data.  After a player
has been assigned to a table, set the tableid to the OpenGameForUser
object and then send out updates to each user.