Create new events for forced check/fold actions
When the action on timer counts down to zero, a command is triggered to
forcefully check or fold the player.  Before this change, the standard
folded/check events were used to represent this, but that implies the
player did it by choice.  Using new "force" check/fold events makes the
history of what really happened more explicit and accurate.