Add a new AutoMoveForward event for all-in situations
This event basically let's the system move the hand forward during an
all-in situation, so that common cards and eventually the game end
without any action on changes or human interaction.  This is just the
event side of this.  An outside process manager will need to be created
to actually cause the hand to continue.  Using an outside source allows
us to time things properly instead of rushing through the game at full
CPU speed.