Update react to 15.5
In 15.5, the react folks have decided that using JS classes is the way
to go, so each createClass() usage has been switched over to that
approach.  As part of the process, the initial state setting was switched
over to be in the constructor and the way event handlers were bound has
been adjusted.